// lib/jobModule/jobPost/types/type.ts
export type LanguageProps = {
  language: string;
  proficiency: string;
};
export type AllLanguageProps = {
  id: string;
  code: string;
  name: string;
  nativeName: string;
};
export type jobCategoryProps = {
  value: string;
  id: string;
};
export type jobSubCategoryProps = {
  value: string;
  id: string;
};
export type jobSkillsProps = {
  value: string;
  id: string;
};